kvtenant: move kv tenant connector out of ccl
This change moves the kv tenant connector out of `kvtenantccl` to enable wider
test coverage of multi-tenant outside `ccl`. There are now two versions of the
kv tenant connector factory, one that provides a connector that only connects
clients to a loopback address, and another that will allow both loopback and
remote addresses. If the loopback version is used in the wrong context an error
will be returned explaining license requirements.

To assist with testing, utilities have been added to instantiate the required kv
connector factory. Tests requiring a kv connector that can connect to remote
nodes can use a testing only factory that is available outside `ccl`. Existing
tests that fell into this category, and depended on `ccl`, have been updated to
use the appropriate utility and connector.

This change does not yet remove the license check for maybe starting a tenant
when a test server is started, since this will most likely break various
existing tests and will require a larger change to fix, it will be done
separately.

Additionally, there are a few other places where `TenantKVAddrs` is populated
with a loopback address, and we can rather favour the new `TenantLoopbackAddr`
to better let the connector factory know that it's only making a loopback
connection and choose the correct logic accordingly. This will be implemented as
part of the test changes mentioned above.

Resolves: #98226
Epic: CRDB-16091
