+++ pkg/BUILD.bazel
@@ -42,7 +42,6 @@ ALL_TESTS = [
     "//pkg/ccl/kvccl/kvfollowerreadsccl:kvfollowerreadsccl_test",
     "//pkg/ccl/kvccl/kvtenantccl/upgradeccl:upgradeccl_test",
     "//pkg/ccl/kvccl/kvtenantccl/upgradeinterlockccl:upgradeinterlockccl_test",
-    "//pkg/ccl/kvccl/kvtenantccl:kvtenantccl_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant-multiregion:3node-tenant-multiregion_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant:3node-tenant_test",
     "//pkg/ccl/logictestccl/tests/5node:5node_test",
@@ -188,6 +187,7 @@ ALL_TESTS = [
     "//pkg/kv/kvclient/kvcoord:kvcoord_disallowed_imports_test",
     "//pkg/kv/kvclient/kvcoord:kvcoord_test",
     "//pkg/kv/kvclient/kvstreamer:kvstreamer_test",
+    "//pkg/kv/kvclient/kvtenant:kvtenant_test",
     "//pkg/kv/kvclient/rangecache:rangecache_test",
     "//pkg/kv/kvclient/rangefeed/rangefeedbuffer:rangefeedbuffer_test",
     "//pkg/kv/kvclient/rangefeed/rangefeedcache:rangefeedcache_test",
@@ -785,7 +785,6 @@ GO_TARGETS = [
     "//pkg/ccl/kvccl/kvtenantccl/upgradeccl:upgradeccl_test",
     "//pkg/ccl/kvccl/kvtenantccl/upgradeinterlockccl:upgradeinterlockccl_test",
     "//pkg/ccl/kvccl/kvtenantccl:kvtenantccl",
-    "//pkg/ccl/kvccl/kvtenantccl:kvtenantccl_test",
     "//pkg/ccl/kvccl:kvccl",
     "//pkg/ccl/logictestccl/tests/3node-tenant-multiregion:3node-tenant-multiregion_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant:3node-tenant_test",
@@ -1203,6 +1202,7 @@ GO_TARGETS = [
     "//pkg/kv/kvclient/kvstreamer:kvstreamer",
     "//pkg/kv/kvclient/kvstreamer:kvstreamer_test",
     "//pkg/kv/kvclient/kvtenant:kvtenant",
+    "//pkg/kv/kvclient/kvtenant:kvtenant_test",
     "//pkg/kv/kvclient/rangecache/rangecachemock:rangecachemock",
     "//pkg/kv/kvclient/rangecache:rangecache",
     "//pkg/kv/kvclient/rangecache:rangecache_test",
