+++ pkg/ccl/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//pkg/ccl/gssapiccl",
         "//pkg/ccl/jwtauthccl",
         "//pkg/ccl/kvccl",
+        "//pkg/ccl/kvccl/kvtenantccl",
         "//pkg/ccl/multiregionccl",
         "//pkg/ccl/multitenantccl",
         "//pkg/ccl/oidcccl",
