+++ pkg/ccl/kvccl/kvtenantccl/BUILD.bazel
@@ -1,108 +1,15 @@
 load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "kvtenantccl",
     srcs = [
-        "connector.go",
-        "rangedesc_iter.go",
-        "setting_overrides.go",
+        "connector_factory.go",
+        "test_utils.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/ccl/kvccl/kvtenantccl",
     visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/config",
-        "//pkg/config/zonepb",
-        "//pkg/gossip",
-        "//pkg/kv",
-        "//pkg/kv/kvclient/kvcoord",
-        "//pkg/kv/kvclient/kvtenant",
-        "//pkg/kv/kvclient/rangecache",
-        "//pkg/kv/kvpb",
-        "//pkg/roachpb",
-        "//pkg/rpc",
-        "//pkg/server/serverpb",
-        "//pkg/settings",
-        "//pkg/spanconfig",
-        "//pkg/ts/tspb",
-        "//pkg/util/contextutil",
-        "//pkg/util/errorutil",
-        "//pkg/util/grpcutil",
-        "//pkg/util/hlc",
-        "//pkg/util/log",
-        "//pkg/util/rangedesc",
-        "//pkg/util/retry",
-        "//pkg/util/syncutil",
-        "//pkg/util/syncutil/singleflight",
-        "//pkg/util/uuid",
-        "@com_github_cockroachdb_errors//:errors",
-        "@com_github_cockroachdb_errors//errorspb",
-        "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//codes",
-        "@org_golang_google_grpc//status",
-    ],
-)
-
-go_test(
-    name = "kvtenantccl_test",
-    size = "large",
-    srcs = [
-        "connector_test.go",
-        "main_test.go",
-        "setting_overrides_test.go",
-        "tenant_kv_test.go",
-        "tenant_range_lookup_test.go",
-        "tenant_scan_range_descriptors_test.go",
-        "tenant_trace_test.go",
-    ],
-    args = ["-test.timeout=895s"],
-    embed = [":kvtenantccl"],
-    shard_count = 16,
-    tags = ["ccl_test"],
-    deps = [
-        "//pkg/base",
-        "//pkg/ccl",
-        "//pkg/config",
-        "//pkg/gossip",
-        "//pkg/keys",
-        "//pkg/kv/kvclient/kvcoord",
-        "//pkg/kv/kvclient/kvtenant",
-        "//pkg/kv/kvclient/rangecache",
-        "//pkg/kv/kvpb",
-        "//pkg/kv/kvserver",
-        "//pkg/kv/kvserver/kvserverbase",
-        "//pkg/kv/kvserver/load",
-        "//pkg/roachpb",
-        "//pkg/rpc",
-        "//pkg/security",
-        "//pkg/security/securityassets",
-        "//pkg/security/securitytest",
-        "//pkg/server",
-        "//pkg/settings",
-        "//pkg/sql",
-        "//pkg/testutils",
-        "//pkg/testutils/serverutils",
-        "//pkg/testutils/sqlutils",
-        "//pkg/testutils/testcluster",
-        "//pkg/util",
-        "//pkg/util/grpcutil",
-        "//pkg/util/hlc",
-        "//pkg/util/leaktest",
-        "//pkg/util/log",
-        "//pkg/util/netutil",
-        "//pkg/util/protoutil",
-        "//pkg/util/randutil",
-        "//pkg/util/rangedesc",
-        "//pkg/util/retry",
-        "//pkg/util/stop",
-        "//pkg/util/tracing/tracingpb",
-        "//pkg/util/uuid",
-        "@com_github_cockroachdb_redact//:redact",
-        "@com_github_stretchr_testify//assert",
-        "@com_github_stretchr_testify//require",
-        "@org_golang_google_grpc//codes",
-        "@org_golang_google_grpc//status",
-    ],
+    deps = ["//pkg/kv/kvclient/kvtenant"],
 )
 
 get_x_data(name = "get_x_data")
