+++ pkg/ccl/sqlproxyccl/tenant/BUILD.bazel
@@ -59,7 +59,6 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/ccl",
-        "//pkg/ccl/kvccl/kvtenantccl",
         "//pkg/ccl/sqlproxyccl/tenantdirsvr",
         "//pkg/roachpb",
         "//pkg/security/securityassets",
