+++ pkg/ccl/testccl/sqlccl/run_control_test.go
@@ -17,7 +17,6 @@ import (
 	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/base"
-	"github.com/cockroachdb/cockroach/pkg/ccl/kvccl/kvtenantccl"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqltestutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/sqlutils"
@@ -63,9 +62,6 @@ func makeRunControlTestCases(t *testing.T) ([]runControlTestCase, func()) {
 		tc.Stopper().Stop(context.Background())
 	}
 }
-
-// Dummy import to pull in kvtenantccl. This allows us to start tenants.
-var _ = kvtenantccl.Connector{}
 
 func TestCancelSelectQuery(t *testing.T) {
 	defer leaktest.AfterTest(t)()
