+++ pkg/cli/BUILD.bazel
@@ -348,7 +348,6 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/build",
-        "//pkg/ccl/kvccl/kvtenantccl",
         "//pkg/cli/clicfg",
         "//pkg/cli/clierror",
         "//pkg/cli/clierrorplus",
@@ -364,6 +363,7 @@ go_test(
         "//pkg/jobs/jobspb",
         "//pkg/keys",
         "//pkg/kv",
+        "//pkg/kv/kvclient/kvtenant",
         "//pkg/kv/kvpb",
         "//pkg/kv/kvserver",
         "//pkg/kv/kvserver/allocator/storepool",
