+++ pkg/kv/kvclient/kvtenant/rangedesc_iter.go
@@ -1,17 +1,19 @@
-// Copyright 2022 The Cockroach Authors.
+// Copyright 2023 The Cockroach Authors.
 //
-// Licensed as a CockroachDB Enterprise file under the Cockroach Community
-// License (the "License"); you may not use this file except in compliance with
-// the License. You may obtain a copy of the License at
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
 //
-//     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
 
-package kvtenantccl
+package kvtenant
 
 import "github.com/cockroachdb/cockroach/pkg/roachpb"
 
 // rangeDescIterator is the concrete (private) implementation of the
-// rangedesc.Iterator interface used by the Connector.
+// rangedesc.Iterator interface used by the connector.
 type rangeDescIterator struct {
 	rangeDescs []roachpb.RangeDescriptor
 	curIdx     int
