+++ pkg/kv/kvclient/rangefeed/rangefeedcache/BUILD.bazel
@@ -36,8 +36,8 @@ go_test(
     embed = [":rangefeedcache"],
     deps = [
         "//pkg/base",
-        "//pkg/ccl/kvccl/kvtenantccl",
         "//pkg/kv",
+        "//pkg/kv/kvclient/kvtenant",
         "//pkg/kv/kvclient/rangecache",  # keep
         "//pkg/kv/kvclient/rangefeed",
         "//pkg/kv/kvclient/rangefeed/rangefeedbuffer",
