+++ pkg/server/BUILD.bazel
@@ -444,7 +444,6 @@ go_test(
         "//pkg/base",
         "//pkg/base/serverident",
         "//pkg/build",
-        "//pkg/ccl/kvccl/kvtenantccl",
         "//pkg/cli/exit",
         "//pkg/clusterversion",
         "//pkg/config",
@@ -454,6 +453,7 @@ go_test(
         "//pkg/jobs/jobspb",
         "//pkg/keys",
         "//pkg/kv",
+        "//pkg/kv/kvclient/kvtenant",
         "//pkg/kv/kvpb",
         "//pkg/kv/kvserver",
         "//pkg/kv/kvserver/allocator",
