+++ pkg/sql/telemetry_test.go
@@ -17,7 +17,6 @@ import (
 	"testing"
 
 	"github.com/cockroachdb/cockroach/pkg/base"
-	"github.com/cockroachdb/cockroach/pkg/ccl/kvccl/kvtenantccl"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/catconstants"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqltestutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
@@ -27,10 +26,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/stretchr/testify/require"
 )
-
-// Dummy import to pull in kvtenantccl. This allows us to start tenants.
-// TODO(yuzefovich): break up the dependency on CCL code.
-var _ = kvtenantccl.Connector{}
 
 func TestTelemetry(t *testing.T) {
 	defer leaktest.AfterTest(t)()
